[refactor] refactor types for food: add type CommonColumn and modify recipe, ingredient type
