+++ backend/src/types/common.ts
@@ -1,0 +1,8 @@
+import { RowDataPacket } from 'mysql2';
+
+export interface CommonColumn extends RowDataPacket {
+  id: number;
+  createTime: Date;
+  updateTime: Date;
+  active: boolean;
+}
